\name{betaplasma}
\alias{betaplasma}
\docType{data}
\title{
Betaplasma dataset
}
\description{Example dataset}
\usage{data(betaplasma)}
\format{
  A data frame with 315 observations on the following 15 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{sex}}{a factor with levels \code{F} \code{M}}
    \item{\code{smokestat}}{a factor with levels \code{Never} \code{Former} \code{Current}}
    \item{\code{bmi}}{a numeric vector}
    \item{\code{vituse}}{a factor with levels \code{Often} \code{Not often} \code{No}}
    \item{\code{calories}}{a numeric vector}
    \item{\code{fat}}{a numeric vector}
    \item{\code{fiber}}{a numeric vector}
    \item{\code{alcohol}}{a numeric vector}
    \item{\code{chol}}{a numeric vector}
    \item{\code{betadiet}}{a numeric vector}
    \item{\code{retdiet}}{a numeric vector}
    \item{\code{betacaro}}{a numeric vector}
    \item{\code{retplasma}}{a numeric vector}
  }
}
\details{
Unknown
}
\source{
Unknown
}
\references{
unknown
}
\examples{
data(betaplasma)
}
\keyword{datasets}
